+++ com.unity.render-pipelines.core/Editor/ShaderStripping/ShaderStrippingReport.cs
@@ -1,0 +1,345 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
+using System.Linq;
+using System.Text;
+using JetBrains.Annotations;
+using UnityEditor.Build;
+using UnityEditor.Build.Reporting;
+using UnityEngine;
+using UnityEngine.Rendering;
+
+namespace UnityEditor.Rendering
+{
+    #region Internal
+    [Serializable]
+    class VariantCounter
+    {
+        public uint inputVariants;
+        public uint outputVariants;
+        public override string ToString() => $"Total={inputVariants}/{outputVariants}({outputVariants / (float)inputVariants * 100f:0.00}%)";
+    }
+
+    [Serializable]
+    class ShaderVariantInfo : VariantCounter
+    {
+        public string variantName;
+        public double stripTimeMs;
+        public override string ToString() => $"{variantName} - {base.ToString()} - Time={stripTimeMs}ms";
+    }
+
+    [Serializable]
+    class ShaderStrippingInfo : VariantCounter, ISerializationCallbackReceiver
+    {
+        public string name;
+
+        private Dictionary<string, List<ShaderVariantInfo>> m_VariantsByPipeline = new();
+
+        public void AddVariant(string pipeline, ShaderVariantInfo variant)
+        {
+            if (!m_VariantsByPipeline.TryGetValue(pipeline, out var list))
+            {
+                list = new List<ShaderVariantInfo>();
+                m_VariantsByPipeline.Add(pipeline, list);
+            }
+
+            inputVariants += variant.inputVariants;
+            outputVariants += variant.outputVariants;
+
+            list.Add(variant);
+        }
+
+        public override string ToString() => $"{name} - {base.ToString()}";
+
+        public void Log(ShaderVariantLogLevel logLevel)
+        {
+            IEnumerable<ShaderVariantInfo> variantsToLog = null;
+            switch (logLevel)
+            {
+                case ShaderVariantLogLevel.AllShaders:
+                {
+                    variantsToLog = m_VariantsByPipeline.SelectMany(i => i.Value);
+                    break;
+                }
+                case ShaderVariantLogLevel.OnlySRPShaders:
+                {
+                    variantsToLog = m_VariantsByPipeline
+                        .Where(i => !string.IsNullOrEmpty(i.Key))
+                        .SelectMany(i => i.Value);
+                    break;
+                }
+            }
+
+            if (variantsToLog != null && variantsToLog.Any())
+            {
+                StringBuilder sb = new StringBuilder();
+                sb.AppendLine($"{this}");
+                foreach (var variant in variantsToLog)
+                {
+                    sb.AppendLine($"- {variant}");
+                }
+
+                Debug.Log(sb.ToString());
+            }
+        }
+
+        [Serializable]
+        class PipelineVariants
+        {
+            public string pipeline;
+            public ShaderVariantInfo[] variants;
+        }
+
+        [SerializeField] private PipelineVariants[] pipelines;
+        public void OnBeforeSerialize()
+        {
+            pipelines = m_VariantsByPipeline
+                .Select(pipeline => new PipelineVariants() { pipeline = pipeline.Key, variants = pipeline.Value.ToArray() })
+                .ToArray();
+        }
+
+        public void OnAfterDeserialize()
+        {
+            pipelines = null;
+        }
+    }
+
+    /// <summary>
+    /// This class works as an scope of the <see cref="ShaderStrippingReport"/> hooking into the
+    /// <see cref="IPreprocessBuildWithReport"/> that are being called at the begin of the build and
+    /// to <see cref="IPostprocessBuildWithReport"/> that are the ones called after the build is finished
+    /// </summary>
+    class ShaderStrippingReportScope : IPostprocessBuildWithReport, IPreprocessBuildWithReport
+    {
+        internal static bool s_DefaultExport = false; // This variable is used by reflection by unit tests
+
+        public int callbackOrder => 0;
+
+        public void OnPreprocessBuild(BuildReport report)
+        {
+            ShaderVariantLogLevel logStrippedVariants = ShaderVariantLogLevel.Disabled;
+            bool exportStrippedVariants = s_DefaultExport;
+
+            // Check the current pipeline and check the shader variant settings
+            if (RenderPipelineManager.currentPipeline != null && RenderPipelineManager.currentPipeline.defaultSettings is IShaderVariantSettings shaderVariantSettings)
+            {
+                logStrippedVariants = shaderVariantSettings.shaderVariantLogLevel;
+                exportStrippedVariants = shaderVariantSettings.exportShaderVariants;
+            }
+
+            ShaderStripping.reporter = (logStrippedVariants == ShaderVariantLogLevel.Disabled && exportStrippedVariants == false) ?
+                new ShaderStrippingReportEmpty() : new ShaderStrippingReport(logStrippedVariants, exportStrippedVariants);
+        }
+
+        public void OnPostprocessBuild(BuildReport report)
+        {
+            ShaderStripping.reporter.DumpReport();
+            ShaderStripping.reporter = null;
+        }
+    }
+
+    /// <summary>
+    /// This class is instantiated as the reporter if the logging and exporting is disabled
+    /// this avoid tracking all the variants, allocating memory, and doing work that is not need
+    /// </summary>
+    class ShaderStrippingReportEmpty : IShaderStrippingReport
+    {
+        public void OnShaderProcessed<TShader, TShaderVariant>([DisallowNull] TShader shader, TShaderVariant shaderVariant, uint variantsIn, uint variantsOut, double stripTimeMs)
+            where TShader : UnityEngine.Object {}
+        public void DumpReport() {}
+    }
+
+    /// <summary>
+    /// Class to gather all the information about stripping in SRP packages
+    /// </summary>
+    class ShaderStrippingReport : IShaderStrippingReport
+    {
+        private readonly ShaderVariantLogLevel m_LogStrippedVariants;
+        private readonly bool m_ExportStrippedVariants;
+
+        // Shader
+        private readonly List<ShaderStrippingInfo> m_ShaderInfos = new();
+        private readonly VariantCounter m_ShaderVariantCounter = new();
+
+        // Compute Shader
+        private readonly List<ShaderStrippingInfo> m_ComputeShaderInfos = new();
+        private readonly VariantCounter m_ComputeShaderVariantCounter = new();
+
+        public ShaderStrippingReport(ShaderVariantLogLevel logLevel, bool export)
+        {
+            m_LogStrippedVariants = logLevel;
+            m_ExportStrippedVariants = export;
+        }
+
+        public void OnShaderProcessed<TShader, TShaderVariant>([DisallowNull] TShader shader, TShaderVariant shaderVariant, uint variantsIn, uint variantsOut, double stripTimeMs)
+            where TShader : UnityEngine.Object
+        {
+            if (!TryGetPipelineAndVariantName(shader, shaderVariant, out string pipeline, out string variantName))
+                throw new NotImplementedException($"Report is not enabled for {typeof(TShader)} and {typeof(TShaderVariant)}");
+
+            var lastShaderStrippingInfo = FindLastShaderStrippingInfo(shader);
+            if (typeof(TShader) == typeof(Shader))
+            {
+                m_ShaderVariantCounter.inputVariants += variantsIn;
+                m_ShaderVariantCounter.outputVariants += variantsOut;
+            }
+            else if (typeof(TShader) == typeof(ComputeShader))
+            {
+                m_ComputeShaderVariantCounter.inputVariants += variantsIn;
+                m_ComputeShaderVariantCounter.outputVariants += variantsOut;
+            }
+
+            lastShaderStrippingInfo.AddVariant(pipeline, new ShaderVariantInfo()
+            {
+                inputVariants = variantsIn,
+                outputVariants = variantsOut,
+                stripTimeMs = stripTimeMs,
+                variantName = variantName
+            });
+        }
+
+        public void DumpReport()
+        {
+            if (m_LogStrippedVariants != ShaderVariantLogLevel.Disabled)
+            {
+                Debug.Log($"Shader Stripping - {m_ShaderVariantCounter}");
+                foreach (var info in m_ShaderInfos)
+                {
+                    info.Log(m_LogStrippedVariants);
+                }
+
+                Debug.Log($"Compute Shader Stripping - {m_ComputeShaderVariantCounter}");
+                foreach (var info in m_ComputeShaderInfos)
+                {
+                    info.Log(m_LogStrippedVariants);
+                }
+            }
+
+            if (m_ExportStrippedVariants)
+            {
+                ExportShaderStrippingInfo("Temp/shader-stripping.json", m_ShaderVariantCounter, m_ShaderInfos);
+                ExportShaderStrippingInfo("Temp/compute-shader-stripping.json", m_ComputeShaderVariantCounter, m_ComputeShaderInfos);
+            }
+        }
+
+        [CanBeNull] private ShaderStrippingInfo m_LastShaderStrippingInfo = null;
+
+        private ShaderStrippingInfo FindLastShaderStrippingInfo<TShader>([DisallowNull] TShader shader)
+            where TShader : UnityEngine.Object
+        {
+            if (m_LastShaderStrippingInfo != null && m_LastShaderStrippingInfo.name.Equals(shader.name))
+                return m_LastShaderStrippingInfo;
+
+            // We are reporting a new shader variant, need to create a new one
+            m_LastShaderStrippingInfo = new ShaderStrippingInfo()
+            {
+                name = shader.name
+            };
+
+            // The compiler will strip the branch that we are not using
+            if (typeof(TShader) == typeof(Shader))
+            {
+                m_ShaderInfos.Add(m_LastShaderStrippingInfo);
+            }
+            else if (typeof(TShader) == typeof(ComputeShader))
+            {
+                m_ComputeShaderInfos.Add(m_LastShaderStrippingInfo);
+            }
+
+            return m_LastShaderStrippingInfo;
+        }
+
+        [MustUseReturnValue]
+        private static bool TryGetPipelineAndVariantName<TShader, TShaderVariant>([DisallowNull] TShader shader, TShaderVariant shaderVariant, out string pipeline, out string variantName)
+            where TShader : UnityEngine.Object
+        {
+            pipeline = string.Empty;
+            variantName = string.Empty;
+
+            if (typeof(TShader) == typeof(Shader) && typeof(TShaderVariant) == typeof(ShaderSnippetData))
+            {
+                var inputShader = (Shader)Convert.ChangeType(shader, typeof(Shader));
+                var snippetData = (ShaderSnippetData)Convert.ChangeType(shaderVariant, typeof(ShaderSnippetData));
+                inputShader.TryGetRenderPipelineTag(snippetData, out pipeline);
+                string passName = string.IsNullOrEmpty(snippetData.passName) ? $"Pass {snippetData.pass.PassIndex}" : snippetData.passName;
+                variantName = $"{passName} ({snippetData.passType}) (SubShader: {snippetData.pass.SubshaderIndex}) (ShaderType: {snippetData.shaderType})";
+            }
+            else if (typeof(TShader) == typeof(ComputeShader) && typeof(TShaderVariant) == typeof(string))
+            {
+                variantName = $"Kernel: {shaderVariant}";
+            }
+            else
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        [Serializable]
+        class Export
+        {
+            public uint totalVariantsIn;
+            public uint totalVariantsOut;
+            public ShaderStrippingInfo[] shaders;
+        }
+
+        void ExportShaderStrippingInfo(string path, VariantCounter variantCounter, List<ShaderStrippingInfo> shaders)
+        {
+            try
+            {
+                var export = new Export()
+                {
+                    totalVariantsIn = variantCounter.inputVariants,
+                    totalVariantsOut = variantCounter.outputVariants,
+                    shaders = shaders.ToArray()
+                };
+
+                File.WriteAllText(path, JsonUtility.ToJson(export, true));
+            }
+            catch (Exception e)
+            {
+                Debug.LogException(e);
+            }
+        }
+    }
+    #endregion
+
+    /// <summary>
+    /// Interface defining a shader stripping report
+    /// </summary>
+    public interface IShaderStrippingReport
+    {
+        /// <summary>
+        /// Method to report a variant stripping
+        /// </summary>
+        /// <param name="shader">Shader can either be <see cref="Shader"/> or <see cref="ComputeShader"/></param>
+        /// <param name="shaderVariant">A variant for the given shader</param>
+        /// <param name="variantsIn">The input variants that were given to the stripper</param>
+        /// <param name="variantsOut">The output variants that were given to the stripper</param>
+        /// <param name="stripTimeMs">The total stripping time in milliseconds</param>
+        /// <typeparam name="TShader">The type of the shader</typeparam>
+        /// <typeparam name="TShaderVariant">The type of the variant</typeparam>
+        /// <exception cref="NotImplementedException">The exception if the Shader doesn't match it's supported variant</exception>
+        void OnShaderProcessed<TShader, TShaderVariant>([DisallowNull] TShader shader,
+            TShaderVariant shaderVariant, uint variantsIn, uint variantsOut, double stripTimeMs)
+            where TShader : UnityEngine.Object;
+
+        /// <summary>
+        /// Dumps the report
+        /// </summary>
+        void DumpReport();
+    }
+
+    /// <summary>
+    /// Global instance to the reporter
+    /// </summary>
+    public static class ShaderStripping
+    {
+        /// <summary>
+        /// Instance to the current build report for shader stripping
+        /// </summary>
+        public static IShaderStrippingReport reporter { get; internal set; }
+    }
+}