+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Disabled volumetric clouds in lens flares sample indoor scene.
 - Make Vertical gate fit the default for physical camera.
 - Changed how the ambient probe is sent to the volumetric clouds trace pass (case 1381761).
+- Changed the json exported with shader variants information when the option `Export Shader Variants` is enabled on the Global settings.
 - Moved custom Sensor Lidar path tracing code to the SensorSDK package.
 - Optimized real time probe rendering by avoid an unnecessary copy per face.